Move `ImagePreview` to separate file
